+++ packages/data-loader/src/commands/export.ts
@@ -14,6 +14,7 @@ export const builder = (args: yargs.Argv) =>
       default: process.env.KINTONE_BASE_URL,
       defaultDescription: "KINTONE_BASE_URL",
       type: "string",
+      demandOption: true,
     })
     .option("username", {
       alias: "u",
@@ -50,6 +51,7 @@ export const builder = (args: yargs.Argv) =>
     .option("app", {
       describe: "The ID of the app",
       type: "string",
+      demandOption: true,
     })
     .option("guest-space-id", {
       describe: "The ID of guest space",
@@ -66,9 +68,13 @@ export const builder = (args: yargs.Argv) =>
       default: "json" as ExportFileFormat,
       choices: formats,
     })
-    .option("query", {
-      alias: "q",
+    .option("condition", {
+      alias: "c",
       describe: "The query string",
+      type: "string",
+    })
+    .option("order-by", {
+      description: "The sort order as a query",
       type: "string",
     })
     .option("pfx-file-path", {
@@ -78,12 +84,7 @@ export const builder = (args: yargs.Argv) =>
     .option("pfx-file-password", {
       describe: "The password of client certificate file",
       type: "string",
-    })
-    // NOTE: Since yargs doesn't detect the type correctly by adding `demandOption: true` in `option()`,
-    // (inferred type always contains `| undefined`)
-    // related issue: https://github.com/yargs/yargs/issues/1928
-    // we declare the required params later as a workaround.
-    .demandOption(["base-url", "app"]);
+    });
 
 type Args = yargs.Arguments<
   ReturnType<typeof builder> extends yargs.Argv<infer U> ? U : never
@@ -101,7 +102,8 @@ export const handler = (args: Args) => {
     guestSpaceId: args["guest-space-id"],
     attachmentDir: args["attachment-dir"],
     format: args.format,
-    query: args.query,
+    condition: args.condition,
+    orderBy: args["order-by"],
     pfxFilePath: args["pfx-file-path"],
     pfxFilePassword: args["pfx-file-password"],
   });