+++ packages/data-loader/src/commands/import.ts
@@ -12,6 +12,7 @@ export const builder = (args: yargs.Argv) =>
       default: process.env.KINTONE_BASE_URL,
       defaultDescription: "KINTONE_BASE_URL",
       type: "string",
+      demandOption: true,
     })
     .option("username", {
       alias: "u",
@@ -48,6 +49,7 @@ export const builder = (args: yargs.Argv) =>
     .option("app", {
       describe: "The ID of the app",
       type: "string",
+      demandOption: true,
     })
     .option("guest-space-id", {
       describe: "The ID of guest space",
@@ -58,6 +60,7 @@ export const builder = (args: yargs.Argv) =>
     .option("file-path", {
       describe: 'The path to source file. ".json" or ".csv"',
       type: "string",
+      demandOption: true,
     })
     .option("pfx-file-path", {
       describe: "The path to client certificate file",
@@ -66,12 +69,7 @@ export const builder = (args: yargs.Argv) =>
     .option("pfx-file-password", {
       describe: "The password of client certificate file",
       type: "string",
-    })
-    // NOTE: Since yargs doesn't detect the type correctly by adding `demandOption: true` in `option()`,
-    // (inferred type always contains `| undefined`)
-    // related issue: https://github.com/yargs/yargs/issues/1928
-    // we declare the required params later as a workaround.
-    .demandOption(["base-url", "app", "file-path"]);
+    });
 
 type Args = yargs.Arguments<
   ReturnType<typeof builder> extends yargs.Argv<infer U> ? U : never